parson: re-implement append_string() as macro
append_string() uses sprintf(), which is un-needed, and is slightly slower
than a simple memcpy().

Also the APPEND_STRING() macro is called only on string constants, so we
can review where strings would be unsafe (and they all look safe so far).

This has a bit of a performance improvement on slower embedded systems.

It also looks like checking for `written` negative is a moot point.

Signed-off-by: Alexandru Ardelean <ardeleanalex@gmail.com>