parson: allow users to provide custom serializer functions for floats
Luckily there's this article which has some neat graphs, explaining the
bigger picture:
   https://aras-p.info/blog/2022/02/25/Curious-lack-of-sprintf-scaling/

Bottom line is: sprintf() is darn slow with numbers (especially
double/float). In our system (an stm32mp1 derivative, which has a simple
A7), we need to convert a json of about 200 numbers. This takes anywhere
from 50-80 ms, depending on the whole system load (50 ms is when the system
isn't too busy).

After digging into this, we've found sprintf() to be the culprit, and some
searches on the internet provided the article above.

The solution for us is to use another double-to-string function.

Other people (on other forums) have revolted against re-implementing
sprintf() as it is a standard function. So, this looks like a good
mitigation/solution for now, to allow for other custom formatter functions.

Some users could just override this to use itoa(), which assumes that their
JSONs are all ints. This also may create the requirement to pass the
formatter as a parameter on the json_serialize_to_string() function.

For our case, just this custom global formatter override is fine.

Signed-off-by: Alexandru Ardelean <ardeleanalex@gmail.com>